requirements.txt
requests
python-dotenv
streamlit
pandas
plotly
transformers
torch
pytest
sqlparse
